Make `get_user_from_token!` generic
So it can be reused by the provider UI.
